refactor(terminal-workspace): extract settings view
